Missing values functionality for toolgroups + more
- New component `annot-tool-group` allows selecting missing values for
columns in tool groups and keeps track of corresponding subject
availability
- `annot-grouped-values` component renamed to `annot-tool-group` component
- `annot-tab` component now supports requirements for
 new component `annot-tool-group`
- `annot-vocabulary-column` component is removed since it has been replaced
by the new `annot-tool-group`
- `annot-vocabulary-row` is now just named `annot-vocabulary`
- Entry criteria to annotation page from categorization page now includes
the selection of one (and only one) subject
- Instructions for subject-column categorization and tool grouping now
appear near 'next page' button on categorization if criteria are unmet
- Adjustments for spacing with 'next page' button on categorization page
in `categ-tool-group`
- New getter `getColumnOfCategory` in store
- New `isDataAnnotated` criteria in store includes if new columns have been
added to the annotated data table. This is from the new 'available'
columns added to the table by the new `annot-tool-group` component when
determining subject availability for each tool group
- transformHeuristics temporarily commented out
- General cleanup: added comments, removing commented out code, removing
unused values/props, etc.