Re-initialized mocked store for column-linking-table test
I ran into a weird problem where I was not able to wrap the commit
method on the store object in a cypress spy, only when the test
ran after another test.

I assume that running the other test somehow changed the store object.
Re-initializing the mocked store object with beforeEach fixed this.